Rename Interactable_Object script and meta file, Finished the Touch detector.
